# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "Stable, machine-readable code, e.g. `DECISION_NOT_FOUND`.",
            "type": "string"
          },
          "details": {
            "description": "Engine error document (failing node and reason) for evaluation failures."
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            },
            "type": [
              "object",
              "null"
            ]
          },
          "message": {
            "description": "Safe to show to a person.",
            "type": "string"
          },
          "requestId": {
            "description": "Same value as the `x-request-id` response header; quote it when reporting a problem.",
            "type": "string"
          }
        },
        "required": [
          "message",
          "code",
          "requestId"
        ],
        "type": "object"
      },
...
      "IssueLogTokenRequest": {
        "properties": {
          "environment": {
            "$ref": "#/components/schemas/EnvironmentName"
          },
          "name": {
            "description": "1 to 100 characters, e.g. the Runtime's host.",
            "maxLength": 100,
            "type": "string"
          }
        },
        "required": [
          "environment",
          "name"
        ],
        "type": "object"
      },
...
      "LogTokenListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LogTokenResponse"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "RequestApprovalRequest": {
        "properties": {
          "releaseId": {
            "description": "The release live on staging.",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "releaseId"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Control plane of Donka: authoring, simulation, releases.",
    "license": {
      "name": ""
    },
    "title": "Donka Studio API",
    "version": "unknown"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/decision-log/tokens": {
      "get": {
        "operationId": "tokens",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogTokenListResponse"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not an administrator (FORBIDDEN)"
          }
        },
        "summary": "Every decision-log token, live ones first; never their value (administrators).",
        "tags": [
          "decision-log"
        ]
      },
      "post": {
        "operationId": "issue_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueLogTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedLogTokenResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or too long name (INVALID_REQUEST)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not an administrator (FORBIDDEN)"
          }
        },
        "summary": "Issues a token Runtimes of one environment send their records with\n(administrators). The token is in this response only.",
        "tags": [
          "decision-log"
        ]
      }
    },
    "/decision-log/tokens/{token_id}": {
      "delete": {
        "operationId": "revoke_token",
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not an administrator (FORBIDDEN)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unknown or already revoked (TOKEN_NOT_FOUND)"
          }
        },
        "summary": "Revokes a decision-log token: batches sent with it are refused from now on (administrators).",
        "tags": [
          "decision-log"
        ]
      }
    },
...
    "/projects/{project_id}/approvals": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Page size, 1 to 100 (default 50).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalListResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND"
          }
        },
        "summary": "The project's production requests, newest first (any member).",
        "tags": [
          "approvals"
        ]
      },
      "post": {
        "operationId": "request",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Viewers cannot ask (FORBIDDEN)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND or RELEASE_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "APPROVAL_PENDING (another request waits) or PROJECT_ARCHIVED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "NOT_ON_STAGING or NO_APPROVER (no other owner)"
          }
        },
        "summary": "Asks for the release live on staging to go to production; the owners who\nmay decide are emailed (editors and owners).",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/projects/{project_id}/environments": {
      "get": {
        "operationId": "environments",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentListResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND"
          }
        },
        "summary": "Both environments: what is live, what was last asked for, how many tokens (any member).",
        "tags": [
          "environments"
        ]
      }
    },
...
    "/projects/{project_id}/environments/{environment}/deployments/{deployment_id}/retry": {
      "post": {
        "operationId": "retry",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EnvironmentName"
            }
          },
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            },
            "description": "Queued again"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Viewers cannot retry (FORBIDDEN)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND, ENVIRONMENT_NOT_FOUND or DEPLOYMENT_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "NOT_RETRYABLE (it did not give up) or PROJECT_ARCHIVED"
          }
        },
        "summary": "Tries a deployment that gave up again now (editors and owners).",
        "tags": [
          "environments"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}